🚚 Renamed `HomeHeader` layout component and added to index
